Refactor all core/scripting/ classes
