+++ core/assets/PrefabAsset.cc
@@ -9,7 +9,7 @@
 #include "core/components/scriptable/TransformComponent.h"
 #include "core/components/synchronized/MeshRendererComponent.h"
 #include "core/components/synchronized/RigidBodyComponent.h"
-#include "core/scripting/ScriptEnvironment.h"
+#include "core/scripting/environment/ScriptEnvironment.h"
 #include "core/world/World.h"
 
 namespace mondradiko {
