+++ core/assets/ScriptAsset.cc
@@ -3,8 +3,8 @@
 
 #include "core/assets/ScriptAsset.h"
 
-#include "core/scripting/ComponentScript.h"
-#include "core/scripting/ScriptEnvironment.h"
+#include "core/scripting/environment/ScriptEnvironment.h"
+#include "core/scripting/instance/ComponentScript.h"
 #include "types/assets/ScriptAsset_generated.h"
 
 namespace mondradiko {
