+++ core/components/scriptable/TransformComponent.h
@@ -4,7 +4,7 @@
 #pragma once
 
 #include "core/components/ScriptableComponent.h"
-#include "core/scripting/ScriptObject.h"
+#include "core/scripting/object/ScriptObject.h"
 #include "core/world/Entity.h"
 #include "lib/include/glm_headers.h"
 #include "types/assets/PrefabAsset_generated.h"
