+++ core/scripting/environment/ScriptEnvironment.cc
@@ -1,14 +1,14 @@
 // Copyright (c) 2020-2021 the Mondradiko contributors.
 // SPDX-License-Identifier: LGPL-3.0-or-later
 
-#include "core/scripting/ScriptEnvironment.h"
+#include "core/scripting/environment/ScriptEnvironment.h"
 
 #include "core/assets/ScriptAsset.h"
 #include "core/components/internal/ScriptComponent.h"
 #include "core/components/scriptable/PointLightComponent.h"
 #include "core/components/scriptable/TransformComponent.h"
-#include "core/scripting/ComponentScript.h"
-#include "core/scripting/ScriptInstance.h"
+#include "core/scripting/instance/ComponentScript.h"
+#include "core/scripting/instance/ScriptInstance.h"
 #include "core/ui/GlyphStyle.h"
 #include "core/ui/UiPanel.h"
 #include "core/world/ScriptEntity.h"
