+++ core/scripting/environment/ScriptEnvironment.h
