+++ core/scripting/instance/ComponentScript.cc
@@ -1,11 +1,11 @@
 // Copyright (c) 2020-2021 the Mondradiko contributors.
 // SPDX-License-Identifier: LGPL-3.0-or-later
 
-#include "core/scripting/ComponentScript.h"
+#include "core/scripting/instance/ComponentScript.h"
 
 #include <array>
 
-#include "core/scripting/ScriptEnvironment.h"
+#include "core/scripting/environment/ScriptEnvironment.h"
 
 namespace mondradiko {
 namespace core {
