+++ core/scripting/instance/UiScript.cc
@@ -1,7 +1,7 @@
 // Copyright (c) 2020-2021 the Mondradiko contributors.
 // SPDX-License-Identifier: LGPL-3.0-or-later
 
-#include "core/scripting/UiScript.h"
+#include "core/scripting/instance/UiScript.h"
 
 #include "core/ui/UiPanel.h"
