+++ core/ui/UserInterface.cc
@@ -14,8 +14,8 @@
 #include "core/gpu/GpuVector.h"
 #include "core/gpu/GraphicsState.h"
 #include "core/renderer/Renderer.h"
-#include "core/scripting/ScriptEnvironment.h"
-#include "core/scripting/UiScript.h"
+#include "core/scripting/environment/ScriptEnvironment.h"
+#include "core/scripting/instance/UiScript.h"
 #include "core/shaders/panel.frag.h"
 #include "core/shaders/panel.vert.h"
 #include "core/ui/GlyphStyle.h"
