+++ core/world/World.h
@@ -5,8 +5,8 @@
 
 #include "core/assets/AssetPool.h"
 #include "core/physics/Physics.h"
-#include "core/scripting/ScriptEnvironment.h"
-#include "core/scripting/ScriptObject.h"
+#include "core/scripting/environment/ScriptEnvironment.h"
+#include "core/scripting/object/ScriptObject.h"
 #include "core/world/Entity.h"
 #include "lib/include/flatbuffers_headers.h"
